+++ sample/app/ui-kendo/kendo-sample-form.model.ts
@@ -11,7 +11,6 @@ import {
     DynamicTextAreaModel,
     DynamicTimePickerModel
 } from "@ng-dynamic-forms/core";
-import { customDateRangeValidator } from "../app.validators";
 
 export const KENDO_SAMPLE_FORM_MODEL = [
 
@@ -19,10 +18,7 @@ export const KENDO_SAMPLE_FORM_MODEL = [
 
             id: "stay",
             validators: {
-                customDateRangeValidator: {
-                    name: customDateRangeValidator.name,
-                    args: null
-                }
+                customDateRangeValidator: null
             },
             errorMessages: {
                 customDateRangeValidator: "Invalid period of time selected"