+++ sample/app/ui-primeng/primeng-sample-form.model.ts
@@ -12,7 +12,6 @@ import {
     DynamicTextAreaModel,
     DynamicTimePickerModel
 } from "@ng-dynamic-forms/core";
-import { customDateRangeValidator } from "../app.validators";
 
 export const PRIME_NG_SAMPLE_FORM_MODEL = [
 
@@ -20,10 +19,7 @@ export const PRIME_NG_SAMPLE_FORM_MODEL = [
 
         id: "stay",
         validators: {
-            customDateRangeValidator: {
-                name: customDateRangeValidator.name,
-                args: null
-            }
+            customDateRangeValidator: null
         },
         errorMessages: {
             customDateRangeValidator: "Invalid period of time selected"